requirements.txt
Django
djangorestframework
Pillow
django-filter
django-crispy-forms
crispy-bootstrap5